+++ src/main/kotlin/org/metalscraps/lostark/item/ItemTier.kt
@@ -1,6 +1,6 @@
 package org.metalscraps.lostark.item
 
-enum class ItemTier(val id: Short, val friendlyName: String) {
+enum class ItemTier(val id: Byte, val friendlyName: String) {
     ALL(0, "전체 티어"),
     TIER1(1, "티어 1"),
     TIER2(2, "티어 2"),
